pyproject.toml
[project]
name = "aki"
version = "0.1.0"
description = "Aki — An agentic system with personality and autonomy"
readme = "README.md"
license = { text = "GPL-3.0" }
requires-python = ">=3.11"
authors = [
    { name = "Pigeon.AI", email = "gggzmz@163.com" }
]
keywords = ["agent", "agentic", "personality", "autonomous", "multi-agent", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.9.0",
    "rich>=13.0",
    "httpx>=0.25.0",
    "openai>=1.0",
    "dashscope>=1.25.12",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tavily-python>=0.3.0",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=1.6.0",
    "pymupdf>=1.23.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pydub>=0.25.1",
    "audioop-lts>=0.2.2; python_version>='3.13'",
    "fastapi>=0.135.1",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
mcp = [
    "mcp>=1.0.0",
]
discord = [
    "discord.py>=2.3.0",
]

[project.scripts]
aki = "aki.cli.main:app"
aki-mcp-server = "aki.mcp.server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aki"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
